Solving dependency hell.
